+++ ui/src/app/enemy-portrait/enemy-portrait.component.ts
@@ -1,0 +1,18 @@
+import { Component, OnInit, Input } from '@angular/core';
+import { Enemy } from '../../../src/model/env/enemies/Enemy';
+
+@Component({
+    selector: 'app-enemy-portrait',
+    templateUrl: './enemy-portrait.component.html',
+    styleUrls: ['./enemy-portrait.component.css']
+})
+export class EnemyPortraitComponent implements OnInit {
+
+    @Input() enemy: Enemy; 
+
+    constructor() { }
+
+    ngOnInit() {
+    }
+
+}